requirements.txt
spacy
pdfplumber
python-docx
openpyxl
pyinstaller
pillow
pytesseract
pdf2image
